requirements.txt
allure-pytest
allure-python-commons
gevent
pytest
pytest-repeat
pytest-rerunfailures
pytest-xdist